+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+- Setting to change allowed traits for Bane material
+- Homebrew materials support
+
 ### Changed
+- Rearranged material data
 - Remove minimum increment on material value
 
 ## [0.1.3] - 2025-08-22